Refactor Socks5Socket to use a generic socket generator
In this commit, a SocketGenerator function has been added to the Socks5Socket structure. The purpose of this function is to generate new sockets when needed, rather than always using TcpStream. Also, changes have been made to the tcp_connect_with_timeout function in util/stream.rs to accept a SocketAddr type, instead of using the ToSocketAddrs trait, and several adjustments were made to accommodate these changes across the codebase.